Initial work on Unicode compatibility (UTF-8).
Initial work on Unicode compatibility (UTF-8).  This includes libucd,
which implements the Unicode Character Database.

(this commit originally also included the following change to MCONFIG

-CFLAGS = -g -O3 -I. -W -Wall -DHAVE_PTHREAD_H
+# MOO is single-threaded, so don't burden us with locking
+CFLAGS = -g -O3 -I. -W -Wall -UHAVE_PTHREAD_H

which clearly should not be propagated back to the shared lib.

And yes, this was originally the first change on this branch,
but since it has been reduced to completely inconsequential
changes, it's now at the end.

--wrog)